+++ app/src/main/java/woowacourse/shopping/di/sgDi/StartInjection.kt
@@ -1,8 +1,12 @@
-package woowacourse.shopping.di.sangoonDi
+package woowacourse.shopping.di.sgDi
 
 object StartInjection {
     operator fun invoke(block: StartInjection.() -> Unit) {
         this.block()
+    }
+
+    inline fun <reified T : Any> single(qualifier: String, instance: Any) {
+        DependencyContainer.add(qualifier, instance)
     }
 
     inline fun <reified T : Any> single(instance: Any) {
